Maptest

+++ app/src/main/java/com/example/ecommerce/model/Order.kt
@@ -2,7 +2,7 @@ package com.example.ecommerce.model
 
 data class Order (
                   val  customerId:String="",
-                  val addressOfCustomerLongitude:Int=0,
-                  val addressOfCustomerLatitude:Int=0,
+                  val addressOfCustomerLongitude:Double=0.0,
+                  val addressOfCustomerLatitude:Double=0.0,
                   val orderState : String=""
                 )
